+++ src/main/java/org/sonar/plugins/delphi/type/factory/TypeFactory.java
@@ -49,6 +49,7 @@ import org.sonar.plugins.delphi.type.Type.TypeType;
 import org.sonar.plugins.delphi.type.Type.VariantType.VariantKind;
 import org.sonar.plugins.delphi.type.factory.DelphiStructType.ImagePart;
 import org.sonar.plugins.delphi.type.intrinsic.IntrinsicType;
+import org.sonar.plugins.delphi.type.parameter.Parameter;
 
 public class TypeFactory {
   private static final CompilerVersion VERSION_4 = CompilerVersion.fromVersionSymbol("VER120");
@@ -335,8 +336,13 @@ public class TypeFactory {
   }
 
   private ProceduralType createProcedural(
-      ProceduralKind kind, List<Type> parameterTypes, Type returnType) {
-    return new DelphiProceduralType(proceduralSize(kind), kind, parameterTypes, returnType);
+      ProceduralKind kind, List<Parameter> parameters, Type returnType) {
+    return createProcedural(kind, parameters, returnType, false);
+  }
+
+  private ProceduralType createProcedural(
+      ProceduralKind kind, List<Parameter> parameters, Type returnType, boolean variadic) {
+    return new DelphiProceduralType(proceduralSize(kind), kind, parameters, returnType, variadic);
   }
 
   public Type untypedType() {
@@ -413,24 +419,28 @@ public class TypeFactory {
     return new DelphiClassReferenceType(type, pointerSize());
   }
 
-  public ProceduralType procedure(List<Type> parameterTypes, Type returnType) {
-    return createProcedural(ProceduralKind.PROCEDURE, parameterTypes, returnType);
+  public ProceduralType procedure(List<Parameter> parameters, Type returnType) {
+    return createProcedural(ProceduralKind.PROCEDURE, parameters, returnType);
   }
 
-  public ProceduralType ofObject(List<Type> parameterTypes, Type returnType) {
-    return createProcedural(ProceduralKind.PROCEDURE_OF_OBJECT, parameterTypes, returnType);
+  public ProceduralType ofObject(List<Parameter> parameters, Type returnType) {
+    return createProcedural(ProceduralKind.PROCEDURE_OF_OBJECT, parameters, returnType);
   }
 
-  public ProceduralType reference(List<Type> parameterTypes, Type returnType) {
-    return createProcedural(ProceduralKind.REFERENCE, parameterTypes, returnType);
+  public ProceduralType reference(List<Parameter> parameters, Type returnType) {
+    return createProcedural(ProceduralKind.REFERENCE, parameters, returnType);
   }
 
-  public ProceduralType anonymous(List<Type> parameterTypes, Type returnType) {
-    return createProcedural(ProceduralKind.ANONYMOUS, parameterTypes, returnType);
+  public ProceduralType anonymous(List<Parameter> parameters, Type returnType) {
+    return createProcedural(ProceduralKind.ANONYMOUS, parameters, returnType);
   }
 
-  public ProceduralType method(List<Type> parameterTypes, Type returnType) {
-    return createProcedural(ProceduralKind.METHOD, parameterTypes, returnType);
+  public ProceduralType method(List<Parameter> parameters, Type returnType) {
+    return createProcedural(ProceduralKind.METHOD, parameters, returnType);
+  }
+
+  public ProceduralType method(List<Parameter> parameters, Type returnType, boolean variadic) {
+    return createProcedural(ProceduralKind.METHOD, parameters, returnType, variadic);
   }
 
   public TypeType typeType(String image, Type type) {