+++ src/main/java/org/sonar/plugins/delphi/type/intrinsic/IntrinsicsInjector.java
@@ -62,182 +62,206 @@ public final class IntrinsicsInjector {
   }
 
   private void buildMethods() {
-    method("Abs").parameters(type(REAL)).returns(type(REAL));
-    method("Abs").parameters(type(INTEGER)).returns(type(INTEGER));
-    method("Abs").parameters(type(INT64)).returns(type(INT64));
-    method("Addr").parameters(typeFactory.untypedType()).returns(typeFactory.untypedPointer());
-    method("Append").parameters(type(TEXT)).returns(type(INTEGER));
-    method("Assert").parameters(type(BOOLEAN), type(UNICODESTRING)).required(1);
+    method("Abs").param(type(REAL)).returns(type(REAL));
+    method("Abs").param(type(INTEGER)).returns(type(INTEGER));
+    method("Abs").param(type(INT64)).returns(type(INT64));
+    method("Addr").varParam(typeFactory.untypedType()).returns(typeFactory.untypedPointer());
+    method("Append").varParam(type(TEXT)).returns(type(INTEGER));
+    method("Assert").varParam(type(BOOLEAN)).param(type(UNICODESTRING)).required(1);
     method("Assign")
-        .parameters(typeFactory.untypedFile(), type(UNICODESTRING), type(WORD))
+        .varParam(typeFactory.untypedFile())
+        .param(type(UNICODESTRING))
+        .param(type(WORD))
         .required(2);
-    method("Assigned").parameters(typeFactory.untypedType()).returns(type(BOOLEAN));
+    method("Assigned").varParam(typeFactory.untypedType()).returns(type(BOOLEAN));
     method("AssignFile")
-        .parameters(typeFactory.untypedFile(), type(UNICODESTRING), type(WORD))
+        .varParam(typeFactory.untypedFile())
+        .param(type(UNICODESTRING))
+        .param(type(WORD))
         .required(2);
     method("AtomicCmpExchange")
-        .parameters(typeFactory.untypedType(), type(INTEGER), type(INTEGER), type(BOOLEAN))
+        .varParam(typeFactory.untypedType())
+        .param(type(INTEGER))
+        .param(type(INTEGER))
+        .outParam(type(BOOLEAN))
         .required(3)
         .returns(type(INTEGER));
     method("AtomicCmpExchange")
-        .parameters(
-            typeFactory.untypedType(),
-            typeFactory.untypedPointer(),
-            typeFactory.untypedPointer(),
-            type(BOOLEAN))
+        .varParam(typeFactory.untypedType())
+        .param(typeFactory.untypedType())
+        .param(typeFactory.untypedType())
+        .outParam(type(BOOLEAN))
         .required(3)
         .returns(typeFactory.untypedPointer());
     method("AtomicDecrement")
-        .parameters(typeFactory.untypedType(), typeFactory.untypedType())
+        .varParam(typeFactory.untypedType())
+        .param(typeFactory.untypedType())
         .required(1)
         .returns(type(INTEGER));
     method("AtomicExchange")
-        .parameters(typeFactory.untypedType(), type(INTEGER))
+        .varParam(typeFactory.untypedType())
+        .param(type(INTEGER))
         .returns(type(INTEGER));
     method("AtomicExchange")
-        .parameters(typeFactory.untypedType(), typeFactory.untypedPointer())
+        .varParam(typeFactory.untypedType())
+        .param(typeFactory.untypedPointer())
         .returns(typeFactory.untypedPointer());
     method("AtomicIncrement")
-        .parameters(typeFactory.untypedType(), typeFactory.untypedType())
+        .varParam(typeFactory.untypedType())
+        .param(typeFactory.untypedType())
         .required(1)
         .returns(type(INTEGER));
     method("BlockRead")
-        .parameters(
-            typeFactory.untypedFile(), typeFactory.untypedType(), type(INTEGER), type(INTEGER))
+        .varParam(typeFactory.untypedFile())
+        .varParam(typeFactory.untypedType())
+        .param(type(INTEGER))
+        .varParam(type(INTEGER))
         .required(3)
         .returns(type(INTEGER));
     method("BlockWrite")
-        .parameters(
-            typeFactory.untypedFile(), typeFactory.untypedType(), type(INTEGER), type(INTEGER))
+        .varParam(typeFactory.untypedFile())
+        .constParam(typeFactory.untypedType())
+        .param(type(INTEGER))
+        .varParam(type(INTEGER))
         .required(3)
         .returns(type(INTEGER));
     method("Break");
-    method("BuiltInArcTan").parameters(type(REAL)).returns(type(EXTENDED));
-    method("BuiltInArcTan2").parameters(type(REAL), type(REAL)).returns(type(EXTENDED));
-    method("BuiltInCos").parameters(type(REAL)).returns(type(EXTENDED));
-    method("BuiltInLn").parameters(type(REAL)).returns(type(EXTENDED));
-    method("BuiltInLnXPlus1").parameters(type(REAL)).returns(type(EXTENDED));
-    method("BuiltInLog10").parameters(type(REAL)).returns(type(EXTENDED));
-    method("BuiltInLog2").parameters(type(REAL)).returns(type(EXTENDED));
-    method("BuiltInSin").parameters(type(REAL)).returns(type(EXTENDED));
-    method("BuiltInSqrt").parameters(type(REAL)).returns(type(EXTENDED));
-    method("BuiltInTan").parameters(type(REAL)).returns(type(EXTENDED));
-    method("Chr").parameters(type(BYTE)).returns(type(CHAR));
-    method("Close").parameters(typeFactory.untypedFile()).returns(type(INTEGER));
-    method("CloseFile").parameters(typeFactory.untypedFile());
+    method("BuiltInArcTan").param(type(REAL)).returns(type(EXTENDED));
+    method("BuiltInArcTan2").param(type(REAL)).param(type(REAL)).returns(type(EXTENDED));
+    method("BuiltInCos").param(type(REAL)).returns(type(EXTENDED));
+    method("BuiltInLn").param(type(REAL)).returns(type(EXTENDED));
+    method("BuiltInLnXPlus1").param(type(REAL)).returns(type(EXTENDED));
+    method("BuiltInLog10").param(type(REAL)).returns(type(EXTENDED));
+    method("BuiltInLog2").param(type(REAL)).returns(type(EXTENDED));
+    method("BuiltInSin").param(type(REAL)).returns(type(EXTENDED));
+    method("BuiltInSqrt").param(type(REAL)).returns(type(EXTENDED));
+    method("BuiltInTan").param(type(REAL)).returns(type(EXTENDED));
+    method("Chr").param(type(BYTE)).returns(type(CHAR));
+    method("Close").varParam(typeFactory.untypedFile()).returns(type(INTEGER));
+    method("CloseFile").varParam(typeFactory.untypedFile());
     method("Concat")
-        .parameters(type(UNICODESTRING), type(UNICODESTRING))
+        .param(type(UNICODESTRING))
+        .param(type(UNICODESTRING))
         .variadic(type(UNICODESTRING))
         .returns(type(UNICODESTRING));
     method("Concat")
-        .parameters(ANY_DYNAMIC_ARRAY, ANY_DYNAMIC_ARRAY)
+        .param(ANY_DYNAMIC_ARRAY)
+        .param(ANY_DYNAMIC_ARRAY)
         .variadic(ANY_DYNAMIC_ARRAY)
         .returns(ANY_DYNAMIC_ARRAY);
     method("Continue");
     method("Copy")
-        .parameters(type(UNICODESTRING), type(INTEGER), type(INTEGER))
+        .param(type(UNICODESTRING))
+        .param(type(INTEGER))
+        .param(type(INTEGER))
         .returns(type(UNICODESTRING));
     method("Copy")
-        .parameters(ANY_DYNAMIC_ARRAY, type(INTEGER), type(INTEGER))
+        .param(ANY_DYNAMIC_ARRAY)
+        .param(type(INTEGER))
+        .param(type(INTEGER))
         .required(1)
         .returns(type(UNICODESTRING));
-    method("Dec").parameters(ANY_ORDINAL, type(INTEGER)).required(1).returns(type(INTEGER));
-    method("Default")
-        .parameters(ANY_CLASS_REFERENCE)
-        .returns(IntrinsicReturnType.classReferenceValue());
-    method("Delete").parameters(type(UNICODESTRING), type(INTEGER), type(INTEGER));
-    method("Delete").parameters(ANY_DYNAMIC_ARRAY, type(INTEGER), type(INTEGER));
-    method("Dispose").parameters(typeFactory.untypedPointer());
-    method("Eof").parameters(typeFactory.untypedFile()).required(0).returns(type(BOOLEAN));
-    method("Eoln").parameters(typeFactory.untypedFile()).required(0).returns(type(BOOLEAN));
-    method("Erase").parameters(typeFactory.untypedFile());
-    method("Exclude").parameters(ANY_SET, ANY_ORDINAL);
-    method("Exit").parameters(typeFactory.untypedType()).required(0);
+    method("Dec").varParam(ANY_ORDINAL).param(type(INTEGER)).required(1).returns(type(INTEGER));
+    method("Default").param(ANY_CLASS_REFERENCE).returns(IntrinsicReturnType.classReferenceValue());
+    method("Delete").varParam(type(UNICODESTRING)).param(type(INTEGER)).param(type(INTEGER));
+    method("Delete").varParam(ANY_DYNAMIC_ARRAY).param(type(INTEGER)).param(type(INTEGER));
+    method("Dispose").varParam(typeFactory.untypedPointer());
+    method("Eof").varParam(typeFactory.untypedFile()).required(0).returns(type(BOOLEAN));
+    method("Eoln").varParam(typeFactory.untypedFile()).required(0).returns(type(BOOLEAN));
+    method("Erase").varParam(typeFactory.untypedFile());
+    method("Exclude").varParam(ANY_SET).param(ANY_ORDINAL);
+    method("Exit").varParam(typeFactory.untypedType()).required(0);
     method("Fail");
-    method("FilePos").parameters(typeFactory.untypedFile()).returns(type(INTEGER));
-    method("FileSize").parameters(typeFactory.untypedFile()).returns(type(INTEGER));
-    method("FillChar").parameters(typeFactory.untypedType(), type(INTEGER), ANY_ORDINAL);
-    method("Finalize").parameters(typeFactory.untypedType(), type(NATIVEUINT)).required(1);
-    method("FreeMem").parameters(typeFactory.untypedPointer(), type(INTEGER)).required(1);
-    method("GetDir").parameters(type(BYTE), type(UNICODESTRING));
-    method("GetMem").parameters(typeFactory.untypedPointer(), type(INTEGER));
-    method("Halt").parameters(type(INTEGER)).required(0);
-    method("HasWeakRef").parameters(ANY_CLASS_REFERENCE).returns(type(BOOLEAN));
-    method("Hi").parameters(type(INTEGER)).returns(type(BYTE));
+    method("FilePos").varParam(typeFactory.untypedFile()).returns(type(INTEGER));
+    method("FileSize").varParam(typeFactory.untypedFile()).returns(type(INTEGER));
+    method("FillChar").varParam(typeFactory.untypedType()).param(type(INTEGER)).param(ANY_ORDINAL);
+    method("Finalize").varParam(typeFactory.untypedType()).param(type(NATIVEUINT)).required(1);
+    method("FreeMem").varParam(typeFactory.untypedPointer()).param(type(INTEGER)).required(1);
+    method("GetDir").param(type(BYTE)).varParam(type(UNICODESTRING));
+    method("GetMem").varParam(typeFactory.untypedPointer()).param(type(INTEGER));
+    method("Halt").param(type(INTEGER)).required(0);
+    method("HasWeakRef").param(ANY_CLASS_REFERENCE).returns(type(BOOLEAN));
+    method("Hi").param(type(INTEGER)).returns(type(BYTE));
     method("High")
-        .parameters(typeFactory.untypedType())
+        .varParam(typeFactory.untypedType())
         .returns(IntrinsicReturnType.high(typeFactory));
-    method("Inc").parameters(ANY_ORDINAL, type(INTEGER)).required(1);
-    method("Include").parameters(ANY_SET, ANY_ORDINAL);
-    method("Initialize").parameters(typeFactory.untypedType(), type(NATIVEINT)).required(1);
-    method("Insert").parameters(type(UNICODESTRING), type(UNICODESTRING), type(INTEGER));
-    method("Insert").parameters(ANY_DYNAMIC_ARRAY, ANY_DYNAMIC_ARRAY, type(INTEGER));
-    method("IsConstValue").parameters(typeFactory.untypedType()).returns(type(BOOLEAN));
-    method("IsManagedType").parameters(ANY_CLASS_REFERENCE).returns(type(BOOLEAN));
-    method("Length").parameters(type(UNICODESTRING)).returns(type(INTEGER));
-    method("Length").parameters(ANY_ARRAY).returns(type(INTEGER));
-    method("Lo").parameters(type(INTEGER)).returns(type(BYTE));
-    method("Low")
-        .parameters(typeFactory.untypedType())
-        .returns(IntrinsicReturnType.low(typeFactory));
+    method("Inc").varParam(ANY_ORDINAL).param(type(INTEGER)).required(1);
+    method("Include").varParam(ANY_SET).param(ANY_ORDINAL);
+    method("Initialize").varParam(typeFactory.untypedType()).param(type(NATIVEINT)).required(1);
+    method("Insert").param(type(UNICODESTRING)).varParam(type(UNICODESTRING)).param(type(INTEGER));
+    method("Insert").param(ANY_DYNAMIC_ARRAY).varParam(ANY_DYNAMIC_ARRAY).param(type(INTEGER));
+    method("IsConstValue").param(typeFactory.untypedType()).returns(type(BOOLEAN));
+    method("IsManagedType").param(ANY_CLASS_REFERENCE).returns(type(BOOLEAN));
+    method("Length").param(type(UNICODESTRING)).returns(type(INTEGER));
+    method("Length").param(ANY_ARRAY).returns(type(INTEGER));
+    method("Lo").param(type(INTEGER)).returns(type(BYTE));
+    method("Low").varParam(typeFactory.untypedType()).returns(IntrinsicReturnType.low(typeFactory));
     method("MemoryBarrier");
     method("MulDivInt64")
-        .parameters(type(INT64), type(INT64), type(INT64), type(INT64))
+        .param(type(INT64))
+        .param(type(INT64))
+        .param(type(INT64))
+        .outParam(type(INT64))
         .required(3)
         .returns(type(INT64));
-    method("New").parameters(typeFactory.untypedPointer());
-    method("Odd").parameters(type(INTEGER)).returns(type(BOOLEAN));
-    method("Ord").parameters(ANY_ORDINAL).returns(type(BYTE));
+    method("New").varParam(typeFactory.untypedPointer());
+    method("Odd").param(type(INTEGER)).returns(type(BOOLEAN));
+    method("Ord").param(ANY_ORDINAL).returns(type(BYTE));
     method("Pi").returns(type(EXTENDED));
-    method("Pred").parameters(ANY_ORDINAL).returns(type(INTEGER));
-    method("Ptr").parameters(type(INTEGER)).returns(typeFactory.untypedPointer());
+    method("Pred").param(ANY_ORDINAL).returns(type(INTEGER));
+    method("Ptr").param(type(INTEGER)).returns(typeFactory.untypedPointer());
     method("Read")
-        .parameters(typeFactory.untypedFile(), typeFactory.untypedType())
+        .varParam(typeFactory.untypedFile())
+        .param(typeFactory.untypedType())
         .variadic(typeFactory.untypedType());
-    method("ReadLn").parameters(typeFactory.untypedFile()).variadic(typeFactory.untypedType());
-    method("ReallocMem").parameters(typeFactory.untypedPointer(), type(INTEGER));
-    method("Rename").parameters(typeFactory.untypedFile(), type(UNICODESTRING));
-    method("Reset").parameters(typeFactory.untypedFile(), type(INTEGER)).required(1);
-    method("Rewrite").parameters(typeFactory.untypedFile(), type(INTEGER)).required(1);
-    method("Round").parameters(type(REAL)).returns(type(INT64));
-    method("RunError").parameters(type(BYTE)).required(0);
-    method("Seek").parameters(typeFactory.untypedFile(), type(INTEGER));
-    method("SeekEof").parameters(type(TEXT)).required(0).returns(type(BOOLEAN));
-    method("SeekEoln").parameters(type(TEXT)).required(0).returns(type(BOOLEAN));
-    method("SetLength").parameters(type(UNICODESTRING), type(INTEGER));
-    method("SetLength").parameters(ANY_DYNAMIC_ARRAY, type(INTEGER)).variadic(type(INTEGER));
-    method("SetString").parameters(type(SHORTSTRING), type(PANSICHAR), type(INTEGER));
-    method("SetString").parameters(type(ANSISTRING), type(PANSICHAR), type(INTEGER));
-    method("SetString").parameters(type(WIDESTRING), type(PCHAR), type(INTEGER));
-    method("SetString").parameters(type(UNICODESTRING), type(PCHAR), type(INTEGER));
+    method("ReadLn").varParam(typeFactory.untypedFile()).variadic(typeFactory.untypedType());
+    method("ReallocMem").varParam(typeFactory.untypedPointer()).param(type(INTEGER));
+    method("Rename").varParam(typeFactory.untypedFile()).param(type(UNICODESTRING));
+    method("Reset").varParam(typeFactory.untypedFile()).param(type(INTEGER)).required(1);
+    method("Rewrite").varParam(typeFactory.untypedFile()).param(type(INTEGER)).required(1);
+    method("Round").param(type(REAL)).returns(type(INT64));
+    method("RunError").param(type(BYTE)).required(0);
+    method("Seek").varParam(typeFactory.untypedFile()).param(type(INTEGER));
+    method("SeekEof").varParam(type(TEXT)).required(0).returns(type(BOOLEAN));
+    method("SeekEoln").varParam(type(TEXT)).required(0).returns(type(BOOLEAN));
+    method("SetLength").varParam(type(UNICODESTRING)).param(type(INTEGER));
+    method("SetLength").varParam(ANY_DYNAMIC_ARRAY).param(type(INTEGER)).variadic(type(INTEGER));
+    method("SetString").varParam(type(SHORTSTRING)).param(type(PANSICHAR)).param(type(INTEGER));
+    method("SetString").varParam(type(ANSISTRING)).param(type(PANSICHAR)).param(type(INTEGER));
+    method("SetString").varParam(type(WIDESTRING)).param(type(PCHAR)).param(type(INTEGER));
+    method("SetString").varParam(type(UNICODESTRING)).param(type(PCHAR)).param(type(INTEGER));
     method("SetTextBuf")
-        .parameters(type(TEXT), typeFactory.untypedType(), type(INTEGER))
+        .varParam(type(TEXT))
+        .varParam(typeFactory.untypedType())
+        .param(type(INTEGER))
         .required(2);
-    method("SizeOf").parameters(typeFactory.untypedType()).returns(type(INTEGER));
-    method("Slice").parameters(ANY_ARRAY, type(INTEGER)).returns(typeFactory.untypedPointer());
-    method("Sqr").parameters(type(EXTENDED)).returns(type(EXTENDED));
-    method("Sqr").parameters(type(INTEGER)).returns(type(INTEGER));
-    method("Sqr").parameters(type(REAL)).returns(type(INT64));
-    method("Str").parameters(typeFactory.untypedType(), type(UNICODESTRING));
-    method("Succ").parameters(ANY_ORDINAL).returns(type(INTEGER));
-    method("Swap").parameters(type(INTEGER)).returns(type(INTEGER));
-    method("Trunc").parameters(type(REAL)).returns(type(INT64));
-    method("Truncate").parameters(typeFactory.untypedFile());
-    method("TypeHandle")
-        .parameters(typeFactory.untypedType())
-        .returns(typeFactory.untypedPointer());
-    method("TypeInfo").parameters(typeFactory.untypedType()).returns(typeFactory.untypedPointer());
-    method("TypeOf").parameters(ANY_OBJECT).returns(typeFactory.untypedPointer());
-    method("Val").parameters(type(UNICODESTRING), typeFactory.untypedType(), type(INTEGER));
-    method("VarArrayRedim").parameters(type(VARIANT), type(INTEGER));
-    method("VarCast").parameters(type(VARIANT), type(VARIANT), type(INTEGER));
-    method("VarClear").parameters(type(VARIANT));
-    method("VarCopy").parameters(type(VARIANT), type(VARIANT));
+    method("SizeOf").param(typeFactory.untypedType()).returns(type(INTEGER));
+    method("Slice").varParam(ANY_ARRAY).param(type(INTEGER)).returns(typeFactory.untypedPointer());
+    method("Sqr").param(type(EXTENDED)).returns(type(EXTENDED));
+    method("Sqr").param(type(INTEGER)).returns(type(INTEGER));
+    method("Sqr").param(type(REAL)).returns(type(INT64));
+    method("Str").constParam(typeFactory.untypedType()).varParam(type(UNICODESTRING));
+    method("Succ").param(ANY_ORDINAL).returns(type(INTEGER));
+    method("Swap").param(type(INTEGER)).returns(type(INTEGER));
+    method("Trunc").param(type(REAL)).returns(type(INT64));
+    method("Truncate").varParam(typeFactory.untypedFile());
+    method("TypeHandle").param(typeFactory.untypedType()).returns(typeFactory.untypedPointer());
+    method("TypeInfo").param(typeFactory.untypedType()).returns(typeFactory.untypedPointer());
+    method("TypeOf").param(ANY_OBJECT).returns(typeFactory.untypedPointer());
+    method("Val")
+        .param(type(UNICODESTRING))
+        .varParam(typeFactory.untypedType())
+        .varParam(type(INTEGER));
+    method("VarArrayRedim").varParam(type(VARIANT)).param(type(INTEGER));
+    method("VarCast").varParam(type(VARIANT)).param(type(VARIANT)).param(type(INTEGER));
+    method("VarClear").varParam(type(VARIANT));
+    method("VarCopy").varParam(type(VARIANT)).param(type(VARIANT));
     method("Write")
-        .parameters(typeFactory.untypedFile(), typeFactory.untypedType())
+        .varParam(typeFactory.untypedFile())
+        .param(typeFactory.untypedType())
         .variadic(typeFactory.untypedType());
-    method("Write").parameters(typeFactory.untypedType()).variadic(typeFactory.untypedType());
-    method("WriteLn").parameters(typeFactory.untypedFile()).variadic(typeFactory.untypedType());
+    method("Write").varParam(typeFactory.untypedType()).variadic(typeFactory.untypedType());
+    method("WriteLn").varParam(typeFactory.untypedFile()).variadic(typeFactory.untypedType());
     method("WriteLn").variadic(typeFactory.untypedType());
   }